Move TextToImageInferenceOutput to shortfin_apps/text_to_image
