+++ shortfin/python/shortfin_apps/text_to_image/TextToImageInferenceOutput.py
